moved reducer to dedicated file and FINALLY got the api response to populate html
